pyproject.toml
[tool.poetry]
name = "kasa"
version = "0.4.0"
description = "NLP tools put together by GhanaNLP Team"
authors = [
    "Paul Azunre <paul.Azunre@gmail.com>",
    "Lawrence <bh-204@hotmail.com>",
    ]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
httpx = "^0.28.1"

[tool.poetry.group.test.dependencies]
pytest = "^6.0.0"
pytest-cov = "^5.0.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.12.0"
pre-commit = "^4.0.1"
isort = "^5.13.2"
flake8 = "^7.1.1"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocstrings = {extras = ["python"], version = "^0.26.2"}
mkdocs-jupyter = "^0.25.1"
markdown-include = "^0.8.1"
mkdocs-material = "^9.5.41"
mkdocs-monorepo-plugin = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
